requirements.txt
requests>=2.31.0
paho-mqtt>=1.6.1
astral>=2.2
python-dateutil>=2.9
